Bug 1849269 - Add SharedWorkerManager::mRemoteWorkerController nullptr checking in SharedWorkerManager::Terminate(). r=dom-worker-reviewers,smaug,asuth
Here comes a possible might cause the crash stack.

    RemoteWorkerController creation fails. If it is Shutdown phase or OOM. The point is ServiceWorkerManager::mRemoteWorkerController is kept as nullptr.

    WorkerManagerCreatedRunnable::mManagerWrapper is the only holder of the ServiceWorkerManager. And it would be released once the runnable finished.

    When releasing WorkerManagerCreatedRunnable::mManagerWrapper, ServiceWorkerManager::UnregisterHolder() is called. In the end SharedWorkerManager::Terminate() is triggered to release the SharedWorker by calling SharedWorkerManager::mRemoteWorkerController::Terminate. However, SharedWorkerManager::mRemoteWorkerController is reasonably a nullptr here, then we meet a de-reference on nullptr.

Differential Revision: https://phabricator.services.mozilla.com/D187269

UltraBlame original commit: f3bb72fd36fc969b5285013cdc0c2516f69bc038

+++ dom/workers/sharedworkers/SharedWorkerManager.cpp
@@ -159,6 +159,12 @@ void SharedWorkerManager::Terminate() {
   MOZ_ASSERT(mActors.IsEmpty());
   MOZ_ASSERT(mHolders.IsEmpty());
 
+  
+  
+  if (!mRemoteWorkerController) {
+    return;
+  }
+
   mRemoteWorkerController->Terminate();
   mRemoteWorkerController = nullptr;
 }
